+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Layout styles in repeatable sections, which contain withRepeatableActions decorator has been fixed (UIDATIMP-502)
 * Implement basic layout for Field Mapping Profile details: MARC Bib from MARC Bib (UIDATIMP-297)
 * Cover Mapping Profile Form with tests including field decorators (UIDATIMP-442)
+* Implement add new row functionality for the Field Mapping Profile details (UIDATIMP-497)
 
 ### Bugs fixed:
 * Fix Item status list in field mapping profiles (UIDATIMP-515)